+++ common/src/main/java/io/github/mortuusars/scholar/client/gui/screen/view/InHandSpreadBookViewScreen.java
@@ -4,10 +4,15 @@ import io.github.mortuusars.scholar.Scholar;
 import io.github.mortuusars.scholar.client.gui.Widgets;
 import io.github.mortuusars.scholar.client.gui.widget.BookmarkButton;
 import io.github.mortuusars.scholar.network.Packets;
-import io.github.mortuusars.scholar.network.packet.server.SetBookmarkC2SP;
+import io.github.mortuusars.scholar.network.packet.serverbound.SetBookmarkC2SP;
+import io.github.mortuusars.scholar.network.packet.serverbound.StartedReadingInHandC2SP;
+import io.github.mortuusars.scholar.network.packet.serverbound.StoppedReadingInHandC2SP;
+import net.minecraft.client.Minecraft;
 import net.minecraft.client.gui.components.Button;
 import net.minecraft.client.gui.components.Tooltip;
 import net.minecraft.network.chat.Component;
+import net.minecraft.sounds.SoundEvents;
+import net.minecraft.sounds.SoundSource;
 import net.minecraft.world.InteractionHand;
 import net.minecraft.world.entity.player.Inventory;
 
@@ -19,6 +24,12 @@ public class InHandSpreadBookViewScreen extends SpreadBookViewScreen {
         super(bookAccess, bookColor);
         this.hand = hand;
         setPage(getBookAccess().getBookmarkedPage());
+        Packets.sendToServer(new StartedReadingInHandC2SP(getBookSlot()));
+        Minecraft.getInstance().level.playSound(player, player, SoundEvents.BOOK_PAGE_TURN, SoundSource.PLAYERS, 1, 1);
+    }
+
+    protected int getBookSlot() {
+        return hand == InteractionHand.MAIN_HAND ? player.getInventory().selected : Inventory.SLOT_OFFHAND;
     }
 
     @Override
@@ -66,5 +77,11 @@ public class InHandSpreadBookViewScreen extends SpreadBookViewScreen {
 
         int slot = hand == InteractionHand.MAIN_HAND ? player.getInventory().selected : Inventory.SLOT_OFFHAND;
         Packets.sendToServer(new SetBookmarkC2SP(slot, newBookmarkedPage));
+    }
+
+    @Override
+    public void onClose() {
+        super.onClose();
+        Packets.sendToServer(new StoppedReadingInHandC2SP(getBookSlot()));
     }
 }