+++ common/src/main/java/io/github/mortuusars/scholar/mixin/reading/PiglinModelMixin.java
@@ -1,0 +1,29 @@
+package io.github.mortuusars.scholar.mixin.reading;
+
+import com.llamalad7.mixinextras.injector.wrapoperation.Operation;
+import com.llamalad7.mixinextras.injector.wrapoperation.WrapOperation;
+import com.llamalad7.mixinextras.sugar.Local;
+import io.github.mortuusars.scholar.client.animation.ReadingPose;
+import net.minecraft.client.model.*;
+import net.minecraft.client.model.geom.ModelPart;
+import net.minecraft.world.entity.Mob;
+import net.minecraft.world.entity.monster.Zombie;
+import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.injection.At;
+
+@Mixin(PiglinModel.class)
+public abstract class PiglinModelMixin<T extends Mob> extends PlayerModel<T> {
+    public PiglinModelMixin(ModelPart root, boolean slim) {
+        super(root, slim);
+    }
+
+    @WrapOperation(method = "setupAnim(Lnet/minecraft/world/entity/Mob;FFFFF)V", at = @At(value = "INVOKE", target = "Lnet/minecraft/client/model/AnimationUtils;animateZombieArms(Lnet/minecraft/client/model/geom/ModelPart;Lnet/minecraft/client/model/geom/ModelPart;ZFF)V"))
+    private void onSetupAnim(ModelPart leftArm, ModelPart rightArm, boolean isAggressive,
+                             float attackTime, float ageInTicks, Operation<Void> original, @Local(argsOnly = true) T entity) {
+        if (ReadingPose.getOpenedBookHand(entity) != null) {
+            return; // Stop zombie arms animation, so it wouldn't override ours.
+        }
+
+        original.call(leftArm, rightArm, isAggressive, ageInTicks, ageInTicks);
+    }
+}