+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 # Changelog
 
 ## UNRELEASED
+- Added book reading animations
+  - Animation will start when a player opens Book and Quill or Written Book, and stops when the book is closed.
+  - This is controlled by `scholar:book_open` component. 
+  - Some humanoid mobs will have this animation as well, if they have opened book in hand.
+  - First-person book model reflects the "opened" book state as well.
+- Added sound to book opening.
 - Author of the book can be changed now.
 - Added bookmark - opens the book at the set page.
 - Added in-world tooltip for books placed on a Lectern.